MDEV-35504 Fix enumeration values intersection
According to the C++ standard, "When = are provided in an enumerator-list,
the values of enumerators are defined by those associated constant-expressions.
For any other enumerator whose definition does not have an =, the associated
value is the value of the previous enumerator plus one".

The last list member with constant expression was `tRPAREN= ')'`, which
corresponds to the value 0x29. So, the following keyword_BKA was assigned 0x2A,
keyword_BNL - 0x2B and so on. However, the earlier member `tCOMMA= ','` had
the value 0x2C, and the same value was automatically assigned to
`keyword_NO_BNL`. `tAT= '@'` also intersected with some of following keywords.

The solution here is to explicitly initialize `keyword_BKA` with a value
that is guaranteed to be larger than any special tokens.