MDEV-19248 Implement MASTER_BIND for CHANGE MASTER
Add CHANGE MASTER TO & SHOW REPLICA STATUS field `Master_Bind` that
configures the `MYSQL_OPT_BIND` of the embedded replication client

The `mariadb` option `MYSQL_OPT_BIND` (`--bind-address` with MDEV-21465)
chooses the network interface (e.g., eternet port)
that connects to the server (the Master in this case).

This is a port of MySQL 5.6.2’s WL#3127.