11.8: MDEV-35897 vector index search allocates too much memory for large ef_search
never estimate that a graph search will visit more nodes than there
are in the graph. In fact, let's reduce the graph size by 30%, it'll
increase the false positive rate of a bloom filter by 2% when
visiting the whole graph, it doesn't affect recall noticeably.

we need to read the shared graph size under a lock. let's store it
in the thread-local unused TABLE::used_stat_records member.