MDEV-36511: Dump basic stats of a table to trace
This feature stores the basic stats of the base tables that are used in
a query, to the optimizer trace. This feature is also controlled by
optimizer_record_context, and is not enabled by default. The stats
such as num_of_records present in the table, indexes if present
then their names, along with the average number of records_per_key
with in each index are dumped to the trace.

The approach taken here is to extend the existing function
store_tables_context_in_trace() and add new dump_stats_to_trace()
to opt_trace_ddl_info.cc Several new tests are added to
opt_trace_store_stats.test